Conflicts:
	application/modules/default/controllers/EventsController.php
	public/css/style.css

+++ application/modules/default/controllers/IdeaController.php
@@ -76,11 +76,15 @@ class IdeaController extends Zend_Controller_Action {
 		$this->view->user_info = $user_info;
 		$this->_helper->layout->setLayout('empty');
 		if (isset($_POST['idea_id'])) {
+			$select = $db->select()
+					->from("idea_items")
+					->where("idea_id = ?", $_POST['idea_id']);
+			$idea = $db->fetchRow($select);
 			$me = -1;
 			if (isset($user_info->user_id) && !empty($user_info->user_id))
 				$me = $user_info->user_id;
 
-			if ($me != -1) {
+			if ($me != -1 && $me != $idea['user_id']) {
 				$db->delete("idea_votes", array("user_id = '" . $me . "'", "idea_id = '" . $_POST['idea_id'] . "'"));
 				switch ($_POST['vote_value']) {
 					case "m":
@@ -176,14 +180,20 @@ class IdeaController extends Zend_Controller_Action {
 				->joinLeft("VIEW_idea_comments_total", "VIEW_idea_comments_total.idea_id=idea_items.idea_id", array("total as comments"));
 		if ($type == "finished") {
 			$select->where("idea_items.idea_status = ?", "1");
+			$select->order("idea_items.idea_finishdate DESC");
 		} else {
-			if ($type != "my") {
-				$select->where("idea_items.idea_status = ?", "0");
+			if ($type == "rejected"){
+				$select->where("idea_items.idea_status = ? OR idea_items.idea_vote_sum <= 0", "2");
+				$select->order("idea_items.idea_posted DESC");
+			} else {
+				if ($type != "my") {
+					$select->where("idea_items.idea_status != ?", "1");
+				}
 			}
 		}
 		if ($type == "top") {
 			$select->order("idea_vote_sum DESC");
-			$select->where("idea_items.idea_vote_sum > 0");
+			$select->where("idea_items.idea_vote_sum > 0 AND idea_items.idea_status = 0");
 		} else {
 			$select->order("idea_posted DESC");
 		}
@@ -191,7 +201,7 @@ class IdeaController extends Zend_Controller_Action {
 			$select->where("idea_items.user_id = ?", $me);
 		}
 		if ($type == "unvoted") {
-			$select->where("idea_votes.vote_value is NULL");
+			$select->where("idea_votes.vote_value is NULL AND idea_items.user_id != ?", $me);
 		}
 		$result = $db->fetchAll($select);
 		foreach ($result as $key => $row) {
@@ -258,6 +268,10 @@ class IdeaController extends Zend_Controller_Action {
 
 	public function listfinishedAction() {
 		$this->_forward("list", null, null, array("type" => "finished"));
+	}
+
+	public function listrejectedAction() {
+		$this->_forward("list", null, null, array("type" => "rejected"));
 	}
 
 	public function viewAction() {
